refactor(workspace): Step 5c — state/connection/sender/server → core (RFC-0001 §5 final)
Why: Adım 5'in final PR'ı (5a #91 + 5b #92 merged). RFC'nin "EN RİSKLİ"
adımı 3 PR'a bölünerek tamamlandı. Bu PR 4 dosyayı core'a mekanik olarak
taşır + birkaç boundary fix yapar; davranış parity korunur (302/302 test
pass), CLAUDE.md kuruluş prensipleri commit-öncesi doğrulandı.

Bu, RFC-0001 Foundation refactor'unun **5/8 adımı** demek — protocol
engine artık tamamen UI/global-state'ten arındırılmış durumda.

## Taşınan dosyalar

* `state.rs` SPLIT — plain struct + impl + tipler core'a (467 LOC),
  STATE static + 11 free-fn shim app'ta kalır (498 → 108 LOC)
* `connection.rs` → `core/connection.rs` (1716 LOC)
* `sender.rs` → `core/sender.rs` (1180 LOC)
* `server.rs` → `core/server.rs` (88 LOC)
* `discovery_types.rs` (NEW core, 31 LOC POD) — `DiscoveredDevice` struct
  burada; mDNS resolve fn'leri (`scan`, `parse`) `app/discovery.rs`'de kalır

## Boundary fix'leri (R2 mitigation)

5c öncesi sender/connection 7 site app-only modül çağırıyordu
(`crate::platform::device_name`, `crate::paths::stats_path`,
`crate::i18n::t`, `crate::discovery::DiscoveredDevice`). Hepsi çözüldü:

### AppState 4 yeni field

Path/default resolution app-side `state::init`'te bir kez yapılır,
AppState taşır:

  - `stats_path: PathBuf`
  - `config_path: PathBuf`     (connection.rs Settings::save için, recon
                                fark etmemişti — agent ek hit)
  - `default_device_name: String`
  - `default_download_dir: PathBuf`

Constructor: `AppState::new(settings, identity_path, stats_path,
config_path, default_device_name, default_download_dir)`. App-side
`state::init`'te `crate::platform::*` + `crate::paths::*` resolve edip
geçirir.

### Sender API extension

  pub async fn send_text(
      device: DiscoveredDevice,
      text: String,
      text_summary: String,    // i18n caller-side (`i18n::t("sender.text_summary")`)
      state: Arc<AppState>,
      ui: Arc<dyn UiPort>,
  ) -> Result<()>

### state::get() → Arc<AppState> parametresi (16 site)

Sender 4 + connection 12 = 16 callsite. Her fn `state: Arc<AppState>`
parametresi alır; body'de `state::get()` lookup yerine `state.method()`
doğrudan. Caller chain (`server::accept_loop`, `main::async_main`) tek
noktada `state::get()` clone'u inject ediyor.

### TransferGuard refactor (PR #91 Gemini #3)

  pub struct TransferGuard {
      state: Arc<AppState>,
      transfer_id: String,
  }
  impl Drop for TransferGuard {
      fn drop(&mut self) {
          self.state.unregister_transfer(&self.transfer_id);
      }
  }

Artık global lookup yok; struct kendi state Arc'ını taşır.

### `set_progress_completed_auto_idle` Weak<AppState>

Method form'a dönerken `Arc::downgrade(&Weak)` ile no-op-on-drop
semantics; auto-idle task `weak.upgrade()` yapamazsa skip. Test teardown
ve gelecekteki testability için daha güvenli.

### PlatformOps trait (UiPort'un eşleniği)

connection.rs'de 3 ayrı `crate::platform::open_url` / `copy_to_clipboard`
çağrısı vardı. UiPort pattern'inin (5b) eşleniği `PlatformOps: Send +
Sync` trait'i kuruldu (2 method); `Arc<dyn PlatformOps>` `accept_loop`
zincirine inject ediliyor. App-side `PlatformAdapter` mevcut
`crate::platform::*` fn'lerini sarar.

### DiscoveredDevice (POD core'a)

`discovery.rs` POD struct kısmı `core/discovery_types.rs`'ye taşındı;
`scan`/`parse` (mDNS-bağımlı) app'ta kalır. discovery.rs `pub use
hekadrop_core::discovery_types::DiscoveredDevice` re-export.

## Cargo.toml

`hekadrop-core/Cargo.toml`:
  - tokio features genişledi: `sync` (Mutex/RwLock async),
    `macros` (tokio::select!), `fs` (file I/O)
  - Yeni: `tokio-util = { version = "0.7", features = ["rt"] }`
    (CancellationToken)

## App-side state.rs (yerinde, içerik değişti)

  pub use hekadrop_core::state::*;        // re-export
  static STATE: OnceLock<Arc<AppState>>;
  pub fn init(settings: Settings) {       // path/default'lar inject edilir
      let _ = STATE.set(AppState::new(
          settings,
          &crate::paths::identity_path(),
          crate::paths::stats_path(),
          crate::paths::config_path(),
          crate::platform::device_name(),
          crate::platform::default_download_dir(),
      ));
  }
  pub fn get() -> Arc<AppState> { ... }
  // 11 free-fn shim (set_listen_port, request_*_window, vs.) — UI-side
  // çağrılar için.

## CLAUDE.md pre-commit invariant scan

| Invariant | Sonuç |
|---|---|
| I-1 (core'da app modül) | 5 doc comment hit, **0 code reference** ✓ |
| I-2 (crate-level allow) | 0 yeni ✓ |
| I-3 (map_err pattern) | 0 yeni ✓ |
| I-6 (yeni global state) | 0 yeni static in core ✓ |

## Davranış paritesi kanıtı

* **Test:** 302 pass / 0 fail (öncesi de 302) — birebir
* **Lock semantikleri:** RwLock, Mutex, AtomicBool, AtomicU64 pattern'ları
  korundu (clear_cancel fast/slow path, rate_limiter pop+push,
  stats save snapshot)
* **Async timing:** tokio::select! biased branch, cancel propagation aynı
* **UI mesajları:** UiAdapter (5b) ve yeni PlatformAdapter `crate::i18n::*`
  + `crate::ui::*` çağrılarını birebir aynı argümanlarla forward eder

## RFC-0001 ilerleme

  ✓ 1 — workspace iskele
  ✓ 2 — hekadrop-proto
  ✓ 3 — hekadrop-core (8 leaf modül)
  ✓ 4 — identity/stats/settings/payload → core (R2 inject pattern)
  ✓ 5a — AppState plain struct + impl
  ✓ 5b — UiPort trait + UiNotification
  ✓ 5c — state/connection/sender/server → core (THIS PR)
  Sıradaki — Adım 6: hekadrop-net (discovery + mdns)
  7 — hekadrop-cli stub
  8 — kapanış (lib.rs sil + surface lock + semver freeze)

## Lokal doğrulama

* cargo build --workspace ✓
* cargo clippy --workspace --all-targets --all-features -- -D warnings ✓
* cargo test --workspace ✓ (302 test)
* cargo fmt ✓
* cargo machete ✓ (0 unused dep)
* typos ✓
