+++ crates/hekadrop-app/src/lib.rs
@@ -43,8 +43,8 @@
 // `tests/*.rs` ve `benches/*.rs` ise `hekadrop::crypto::xxx` formuyla bu
 // shim üzerinden core'a ulaşır.
 pub use hekadrop_core::{
-    config, crypto, error, file_size_guard, frame, identity, log_redact, payload, secure, settings,
-    stats, ui_port, ukey2,
+    config, connection, crypto, discovery_types, error, file_size_guard, frame, identity,
+    log_redact, payload, secure, sender, server, settings, stats, ui_port, ukey2,
 };
 
 // RFC-0001 §5 Adım 2: protobuf bindings `hekadrop-proto` crate'inden
