+++ crates/hekadrop-core/src/server.rs
@@ -1,6 +1,7 @@
-use crate::connection;
+use crate::connection::{self, PlatformOps};
+use crate::state::AppState;
+use crate::ui_port::UiPort;
 use anyhow::Result;
-use hekadrop_core::ui_port::UiPort;
 use std::sync::Arc;
 use tokio::net::TcpListener;
 use tokio::sync::Semaphore;
@@ -48,7 +49,12 @@ pub async fn start_listener() -> Result<TcpListener> {
     }
 }
 
-pub async fn accept_loop(listener: TcpListener, ui: Arc<dyn UiPort>) -> Result<()> {
+pub async fn accept_loop(
+    listener: TcpListener,
+    ui: Arc<dyn UiPort>,
+    state: Arc<AppState>,
+    platform: Arc<dyn PlatformOps>,
+) -> Result<()> {
     let permits = Arc::new(Semaphore::new(MAX_CONCURRENT_CONNECTIONS));
     loop {
         let (socket, addr) = listener.accept().await?;
@@ -73,8 +79,13 @@ pub async fn accept_loop(listener: TcpListener, ui: Arc<dyn UiPort>) -> Result<(
 
         info!("bağlantı: {}", addr);
         let ui_for_conn = Arc::clone(&ui);
+        let state_for_conn = Arc::clone(&state);
+        let platform_for_conn = Arc::clone(&platform);
         tokio::spawn(async move {
-            if let Err(e) = connection::handle(socket, addr, ui_for_conn).await {
+            if let Err(e) =
+                connection::handle(socket, addr, ui_for_conn, state_for_conn, platform_for_conn)
+                    .await
+            {
                 warn!("bağlantı hatası ({}): {:?}", addr, e);
             }
             drop(permit); // connection bittiğinde permit geri döner
